Use ISO 3166-1 alpha-2 country codes
Change the database and PHP handling of countries to use ISO 3166-1
alpha-2 country codes instead of the old numeric ID method.

The new data is also using the complete ISO country list and standard
ISO country names instead of the prior selective list.

Change includes a migration script that can be used to update the DDL
and contents of an older installation.

Bug: 57547
Change-Id: Ib8b4c20f4f280109ab9a5fa9dd117f1976cacaf7