Use PHPMailer to send mail
Replace PHP's built-in mail() function with PHPMailer so that mail can
be sent via an SMTP relay rather than requiring sendmail on the
application server. The initial implementation assumes that an SMTP
relay will be available on localhost, but this is temporary until other
changesets that introduce better configuration management land in
master.

Bug: 57565
Change-Id: I58221cb7c9c42183ab48f81c4c861f2054be68a4