Remove chosen jquery library
The Chosen library was added to the application to make selecting a wiki
domain name easier when there were hundreds of domain names in the wiki
select list. Now that the wiki select list is only the various projects
rather than the complete expansion of communities and languages we can
remove it and use native select widgets instead.

Change-Id: Ifab14f8ce5d2b5b4e12a675d577b95b7a3bc8d2e