Add support for self-service password resets
Add a screen to request a password reset token to be delivered by email
and an associated screen to validate the reset token and allow
a password reset. The recovery tokens issued will automatically expire
on the next password change or in 48 hours if not used. Backport of
I0eb3aa7 from wikimedia/iegreview.

Bug: T90301
Change-Id: I48effdf830064b4a57ada5211f3b92aaae1b12ac