Move custom js to file and tighten CSP rules
Move the inline javascript for session keepalive to a static js file.
This is needed to comply with the content security policy of 'self'
which keeps inline javascript from executing.

The prior update of jQuery and Modernizer versions also allows us to
update the content security policy rules to exclude unsafe-inline
styles. The rules used are fundamentally the same rules used by Striker
with the addition of data: as an image source which is needed by the
inline sprites used in main.css.

Change-Id: I3b4862c78953a2e82b74f9f150aa228d51476ed4